Resolve S10P12A504-765 "Be websocket"

See merge request s10-webmobile1-sub2/S10P12A504!67

+++ backend/src/main/java/com/quiet/onterview/matching/service/MessageService.java
@@ -10,11 +10,11 @@ public class MessageService {
     private final SimpMessagingTemplate simpMessagingTemplate;
     private final String SUB_PREFIX = "/sub/";
 
-    public void announceAll(Long roomId, String message) {
+    public void announceAll(Integer roomId, String message) {
         simpMessagingTemplate.convertAndSend(SUB_PREFIX + roomId, message);
     }
 
-    public void announceToUser(Long roomId, String user, String message) {
+    public void announceToUser(Integer roomId, String user, String message) {
         simpMessagingTemplate.convertAndSendToUser(user, SUB_PREFIX + roomId, message);
     }
 }
